refactor: 코드리뷰 반영 — Feature-based 구조 전환 및 버그 수정
- app/features/shared/ Feature-based 디렉토리 구조로 전환
- apiFetch 빈 body 200 응답 크래시 수정 (content-type 기반 판단)
- apiFetch GET 요청 시 Content-Type 헤더 미전송
- apiFetch API_BASE undefined 가드 추가
- apiFetch 상대 import를 @/ path alias로 변경
- NotFoundPage Link/Button 중첩 제거 (useNavigate 사용)
- Input focus:border-black 누락 보완
- 전체 interactive translate 값 디자인 시스템 기준(4px/6px)으로 통일
- SettingsModal role=dialog, aria-modal=true 추가
