+++ frontend/src/app/App.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
 import { RouterProvider } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { router } from "@/router";
-import { useAuthStore } from "@/stores/useAuthStore";
+import { router } from "@/app/router";
+import { useAuthStore } from "@/shared/stores/useAuthStore";
 
 const queryClient = new QueryClient({
   defaultOptions: {
