+++ frontend/src/app/layout/Header.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Link, NavLink } from "react-router-dom";
-import { useAuthStore } from "@/stores/useAuthStore";
-import { SettingsModal } from "@/components/layout/SettingsModal";
+import { useAuthStore } from "@/shared/stores/useAuthStore";
+import { SettingsModal } from "@/app/layout/SettingsModal";
 
 const NAV_ITEMS = [
   { to: "/", label: "게임" },
@@ -31,7 +31,7 @@ export function Header() {
                     "border-4 border-black dark:border-white px-4 py-1.5 font-bold text-sm transition-all duration-100",
                     isActive
                       ? "bg-black text-white dark:bg-white dark:text-black shadow-none translate-x-0 translate-y-0"
-                      : "bg-white dark:bg-gray-950 text-black dark:text-white shadow-brutal-sm hover:shadow-brutal-hover hover:translate-x-0.5 hover:translate-y-0.5 active:shadow-none active:translate-x-0.75 active:translate-y-0.75",
+                      : "bg-white dark:bg-gray-950 text-black dark:text-white shadow-brutal-sm hover:shadow-brutal-hover hover:translate-x-1 hover:translate-y-1 active:shadow-none active:translate-x-1.5 active:translate-y-1.5",
                   ].join(" ")
                 }
               >
@@ -43,7 +43,7 @@ export function Header() {
           <div className="flex items-center gap-3">
             <button
               onClick={() => setIsSettingsOpen(true)}
-              className="border-4 border-black dark:border-white bg-white dark:bg-gray-950 text-black dark:text-white px-3 py-1.5 text-sm font-bold shadow-brutal-sm transition-all duration-100 hover:shadow-brutal-hover hover:translate-x-0.5 hover:translate-y-0.5 active:shadow-none active:translate-x-0.75 active:translate-y-0.75"
+              className="border-4 border-black dark:border-white bg-white dark:bg-gray-950 text-black dark:text-white px-3 py-1.5 text-sm font-bold shadow-brutal-sm transition-all duration-100 hover:shadow-brutal-hover hover:translate-x-1 hover:translate-y-1 active:shadow-none active:translate-x-1.5 active:translate-y-1.5"
               aria-label="설정"
             >
               설정
@@ -52,14 +52,14 @@ export function Header() {
             {isAuthenticated && user ? (
               <Link
                 to="/mypage"
-                className="border-4 border-black dark:border-white bg-yellow-300 px-4 py-1.5 text-sm font-bold text-black shadow-brutal-sm transition-all duration-100 hover:shadow-brutal-hover hover:translate-x-0.5 hover:translate-y-0.5 active:shadow-none active:translate-x-0.75 active:translate-y-0.75"
+                className="border-4 border-black dark:border-white bg-yellow-300 px-4 py-1.5 text-sm font-bold text-black shadow-brutal-sm transition-all duration-100 hover:shadow-brutal-hover hover:translate-x-1 hover:translate-y-1 active:shadow-none active:translate-x-1.5 active:translate-y-1.5"
               >
                 {user.nickname}
               </Link>
             ) : (
               <Link
                 to="/login"
-                className="border-4 border-black dark:border-white bg-green-400 px-4 py-1.5 text-sm font-bold text-black shadow-brutal-sm transition-all duration-100 hover:shadow-brutal-hover hover:translate-x-0.5 hover:translate-y-0.5 active:shadow-none active:translate-x-0.75 active:translate-y-0.75"
+                className="border-4 border-black dark:border-white bg-green-400 px-4 py-1.5 text-sm font-bold text-black shadow-brutal-sm transition-all duration-100 hover:shadow-brutal-hover hover:translate-x-1 hover:translate-y-1 active:shadow-none active:translate-x-1.5 active:translate-y-1.5"
               >
                 로그인
               </Link>
