+++ frontend/src/shared/ui/Card.tsx
