+++ frontend/src/shared/ui/Input.tsx
@@ -6,7 +6,7 @@ export function Input({ className = "", ...rest }: InputHTMLAttributes<HTMLInput
       className={[
         "border-4 border-black dark:border-white rounded-none bg-white dark:bg-gray-900",
         "text-black dark:text-white px-4 py-3 font-medium text-lg w-full",
-        "focus:outline-none focus:ring-0 placeholder:text-gray-400",
+        "focus:outline-none focus:ring-0 focus:border-black dark:focus:border-white placeholder:text-gray-400",
         className,
       ].join(" ")}
       {...rest}
